Fix svs_disp tests and add test for paxctld
Some sd-svs-disp tests were not running, moving sd-svs-disp tests to their own file resolves.
